check change description
removed need for carousel (researched it a bit and it's not a good move overall. people understand carousels relatively well, but it hides content, gives the user room for error where if they don't swipe directly down they could scroll on the carousel, and a few other issues).

also changed all the button tags to anchor tags with the proper styling and added the proper references to scroll through the page. just need apply here link and test link.